# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de gestion des catégories",
        "description": "Documentation pour l'API de gestion des catégories de votre application.",
        "contact": {
            "email": "votre_email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/historiques/valider/{evenement}": {
            "post": {
                "summary": "Valider la participation de l'utilisateur actuel à un événement",
                "operationId": "3cf72fe796a3da1be29802d24022c032",
                "parameters": [
                    {
                        "name": "Evenement à valide",
                        "in": "path",
                        "description": "ID de l'événement à valide",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Succès de la validation de l'événement"
                    }
                }
            }
        },
        "/api/permissions": {
            "get": {
                "summary": "Récupérer toutes les permissions",
                "operationId": "27bb1606a7b293d1d49f57bb94ae71f7",
                "responses": {
                    "200": {
                        "description": "Liste des permissions"
                    }
                }
            },
            "post": {
                "summary": "Crée nouvelle permission",
                "operationId": "37c918c86b5085b0eb088fa760c7175d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Permission est crée"
                    }
                }
            }
        },
        "/api/permissions/{id}": {
            "put": {
                "summary": "Modifier permission",
                "operationId": "db11d4d00da3ece5b99187df6f1e86e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Permission est mise à jour"
                    }
                }
            },
            "delete": {
                "summary": "Supprimer la permission",
                "operationId": "f0afbd22579524a5bb83b653f7dd18dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Permission est supprimé"
                    }
                }
            }
        },
...
    },
...
}